// lib/dao/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
	{ [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
	{ [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
};
// ...
export type Group = {
	groupId?: Maybe<Scalars['ID']>;
	title: Scalars['String'];
	members?: Maybe<Array<GroupMember>>;
};

export type GroupMember = User | NonUser;
// ...
export type SecretPick = {
	group?: Maybe<Group>;
	group_member_info?: Maybe<GroupMember>;
	secret_pick?: Maybe<Scalars['String']>;
};
// ...
export type UserDbObject = {
	_id: ObjectID;
	first_name?: Maybe<string>;
	last_name?: Maybe<string>;
	email?: Maybe<string>;
	password?: Maybe<string>;
	groups?: Maybe<Array<Group>>;
	user_secret_picks?: Maybe<Array<SecretPick>>;
};